chore(AAP-91340): Split api.ts into domain-specific modules
